Get rid of explicit primitive types casts

+++ rules/src/consts.rs
@@ -1,5 +1,5 @@
-pub(crate) const SECOND: usize = 1;
-pub(crate) const MINUTE: usize = 60 * SECOND;
-pub(crate) const HOUR: usize = 60 * MINUTE;
-pub(crate) const DAY: usize = 24 * HOUR;
-pub(crate) const WEEK: usize = 7 * DAY;
+pub(crate) const SECOND: u32 = 1;
+pub(crate) const MINUTE: u32 = 60 * SECOND;
+pub(crate) const HOUR: u32 = 60 * MINUTE;
+pub(crate) const DAY: u32 = 24 * HOUR;
+pub(crate) const WEEK: u32 = 7 * DAY;
